+++ libminisip/source/subsystem_media/rtp/SDES_EMAIL.cxx
@@ -58,7 +58,7 @@ int SDES_EMAIL::size(){
 //	int npad = 4-((2+cname.length())%4);
 //	if (npad==4)
 //		npad=0;
-	return 2+email.length()/*+npad*/;
+	return 2+(int)email.length()/*+npad*/;
 }
 
 #ifdef DEBUG_OUTPUT
