+++ libminisip/source/subsystem_media/rtp/SDES_NOTE.cxx
@@ -58,7 +58,7 @@ int SDES_NOTE::size(){
 //	int npad = 4-((2+cname.length())%4);
 //	if (npad==4)
 //		npad=0;
-	return 2+note.length()/*+npad*/;
+	return 2+(int)note.length()/*+npad*/;
 }
 
 #ifdef DEBUG_OUTPUT
